// src/types/GeometryState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type GeometryState = { rotate90: number, flipH: boolean, flipV: boolean, straighten: number, perspectiveV: number, perspectiveH: number, perspectiveRotate: number, aspectAdjust: number, scale: number, offsetX: number, offsetY: number, };